Location Status Timer
Essentially when OPCUA_LOCATION is called we check previous location and see if it lines up with new location. If it does then we reset the timer else we cancel the timer, set the location to false and set new location to true. Also, if timer runs out before we receive next location we set the previous location to false since bucket is probably traveling now.